// src/API/PenMessageType.ts
export enum SettingType {
  TimeStamp = 1,
  AutoPowerOffTime = 2,
  PenCapOff = 3,
  AutoPowerOn = 4,
  Beep = 5,
  Hover = 6,
  OfflineData = 7,
  LedColor = 8,
  Sensitivity = 9,
  UsbMode = 10,
  DownSampling = 11,
  BtLocalName = 12,
  FscSensitivity = 13,
  DataTransmissionType = 14,
  BeepAndLight = 16,
  InitDisk = 17,
}

export enum PenTipType {
  Normal = 0,
  Eraser = 1,
}

export enum ErrorType {
  MissingPenUp = 1,
  MissingPenDown = 2,
  InvalidTime = 3,
  MissingPenDownPenMove = 4,
  ImageProcessingError = 5,
  InvalidEventCount = 6,
  MissingPageChange = 7,
  MissingPenMove = 8,
}

export enum ProfileType {
  CreateProfile = 0x01,
  DeleteProfile = 0x02,
  InfoProfile = 0x03,
  WriteProfileValue = 0x11,
  ReadProfileValue = 0x12,
  DeleteProfileValue = 0x13,
}

export enum FirmwareStatusType {
  STATUS_START = 0,
  STATUS_CONTINUE = 1,
  STATUS_END = 2,
  STATUS_ERROR = 3,
}